[MISC] move type_traits/transformation_trait_or to utility
