+++ include/seqan3/core/detail/template_inspection.hpp
@@ -12,9 +12,9 @@
 
 #pragma once
 
-#include <seqan3/core/type_traits/transformation_trait_or.hpp>
-
 #include <seqan3/std/concepts>
+
+#include <seqan3/utility/type_traits/detail/transformation_trait_or.hpp>
 
 namespace seqan3::detail
 {
