+++ include/seqan3/range/views/interleave.hpp
@@ -18,12 +18,12 @@
 
 #include <range/v3/view/chunk.hpp>
 
-#include <seqan3/core/type_traits/transformation_trait_or.hpp>
 #include <seqan3/range/detail/random_access_iterator.hpp>
 #include <seqan3/range/views/detail.hpp>
 #include <seqan3/range/views/join.hpp>
 #include <seqan3/range/views/persist.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
+#include <seqan3/utility/type_traits/detail/transformation_trait_or.hpp>
 #include <seqan3/utility/type_traits/pre.hpp>
 
 namespace seqan3::detail
