+++ include/seqan3/range/views/take_until.hpp
@@ -21,12 +21,12 @@
 #include <seqan3/core/detail/iterator_traits.hpp>
 #include <seqan3/core/semiregular_box.hpp>
 #include <seqan3/core/type_traits/range.hpp>
-#include <seqan3/core/type_traits/transformation_trait_or.hpp>
 #include <seqan3/io/exception.hpp>
 #include <seqan3/range/concept.hpp>
 #include <seqan3/range/views/detail.hpp>
 #include <seqan3/range/detail/inherited_iterator_base.hpp>
 #include <seqan3/range/container/concept.hpp>
+#include <seqan3/utility/type_traits/detail/transformation_trait_or.hpp>
 
 namespace seqan3::detail
 {
