+++ test/unit/core/type_traits/transformation_trait_or_test.cpp
@@ -7,7 +7,7 @@
 
 #include <gtest/gtest.h>
 
-#include <seqan3/core/type_traits/transformation_trait_or.hpp>
+#include <seqan3/utility/type_traits/detail/transformation_trait_or.hpp>
 
 struct A
 {
